Refactor: 음악 종료 함수화

+++ music/MusicPlayer.js
@@ -29,14 +29,14 @@ class MusicPlayer {
         try {
             const tempDir = `./music/temp/${guildId}`;
             fs.mkdirSync(tempDir, { recursive: true });
-            const ytDlpProcess = spawn("yt-dlp", [
+            const songStream = spawn("yt-dlp", [
                 "-f", "bestaudio",
                 "-o", "-",
                 "--no-playlist",
                 queue.songs[0].url
-            ], {env: process.env, cwd: tempDir});
+            ]);
 
-            const resource = createAudioResource(ytDlpProcess.stdout, {inlineVolume: true});
+            const resource = createAudioResource(songStream.stdout, {inlineVolume: true});
             const audioPlayer = createAudioPlayer();
             const connection = getVoiceConnection(guildId);
             
@@ -44,38 +44,38 @@ class MusicPlayer {
             audioPlayer.play(resource);
             MusicChannel.update(guildId);
 
-        audioPlayer.on(AudioPlayerStatus.Idle, () => {
-            ffmpeg.kill();
+            audioPlayer.on(AudioPlayerStatus.Idle, () => {
+                songStream.kill();
 
                 if (queue.loop === 2)
                     queue.songs.unshift(queue.songs[0]);
                 else if (queue.loop === 1) 
                     queue.songs.push(queue.songs[0]);
 
-                queue.songs.shift();
-
-                if (queue.songs.length === 0) {
-                    connection.destroy();
-                    fs.rmSync(tempDir, { recursive: true, force: true });
-                    MusicChannel.update(guildId).then(() => {
-                        ServerQueue.delete(guildId);
-                    });
-                }
-                else this.play(guildId);
+                this.handleSongEnd(guildId, queue, connection, tempDir);
             });
         } catch (err) {
             console.log("음악 재생 중 에러가 발생했습니다 : " + err);
-            queue.songs.shift();
-            if (queue.songs.length > 0) {
-                this.play(guildId);
-            } else {
-                const connection = getVoiceConnection(guildId);
-                connection.destroy();
-                MusicChannel.update(guildId).then(() => {
-                    ServerQueue.delete(guildId);
-                });
-            }
+            this.handleSongEnd(guildId, queue);
         }
+    }
+
+    handleSongEnd(guildId, queue, connection, tempDir) {
+        if (queue.loop === 2)
+            queue.songs.unshift(queue.songs[0]);
+        else if (queue.loop === 1) 
+            queue.songs.push(queue.songs[0]);
+
+        queue.songs.shift();
+
+        if (queue.songs.length === 0) {
+            connection?.destroy();
+            MusicChannel.update(guildId).then(() => {
+                ServerQueue.delete(guildId); // queue 접근 불가
+            });
+            if (tempDir) fs.rmSync(tempDir, { recursive: true, force: true });
+        }
+        else this.play(guildId);
     }
 }
